Create additional ident for virtual sensors
The current virtual sensors uses the same ident as physical sensors. This will
cause some errors in some circumstances. For example: if an orientation sensor
and an accelerometer are both registered with a same SensorListener, they are
using a same ident. If then the orientation sensor is unregistered from that
SensorListener, the orientation sensor implementation will call into the
accelerometer's deactive method with the same ident of accelerometer. Then
accelerometer will be disabled too for that SensorListener.

This patch creates additional ident for each virtual sensors to avoid possible
problems as described above.

Originally from: http://review.cyanogenmod.org/#/c/108636/